refactor(copilot): extract pure copilot domain into the crate (Phase 5 C1)
Move the copilot's pure decision logic out of the binary's
src/domain/ai/copilot.rs into vailabel_copilot::domain, split into
idiomatic submodules:

- routing.rs  — Capability, RoutedIntent, the keyword router + tool gating
- planning.rs — PlanCapability/PlanStep/Plan, the LLM plan parser/validator,
                planner prompts, deterministic route_to_plan fallback
- qa.rs       — QaFinding/ProposedAction/CopilotTurnResult + the QA-diff engine
- labels.rs   — label-suggestion parsing (JSON/CSV/bulleted)
- config.rs   — CopilotLlmConfig (moved from ai/model.rs)

The one binary coupling (crate::value_string) is inlined as a private
helper in qa.rs; the two doc-comment references to binary paths are
reworded. The 11 unit tests move with the logic (binary lib 41->30,
vailabel_copilot 0->11 — net preserved).

Re-export shims keep every call site byte-identical:
- src/domain/ai/copilot.rs is now `pub use vailabel_copilot::domain::*;`
- ai/model.rs re-exports CopilotLlmConfig from the crate

The crate stays FULLY_PURE (serde + serde_json only; no diesel/tauri/
reqwest/inference) so no arch-test change is needed. The LLM/inference
orchestration that consumes this logic stays in AiService for now (ports
+ app service land in C2-C4). Default cargo check clean; cargo test
--workspace green; arch-tests pass; --no-default-features still shows
only the pre-existing 4 ai/service.rs errors.
