+++ apps/studio/src-tauri/crates/copilot/src/lib.rs
@@ -1,11 +1,16 @@
 //! `vailabel-copilot` — the Copilot module.
 //!
-//! The offline AI labeling copilot: chat, tool calling, prompt/session
-//! management, AI actions, and the approval workflow. Phase 1 extracts the pure
-//! request/response DTOs (the chat-turn payload, the approved-action payload,
-//! and the connection-test payload/result) into [`contracts`]. The
-//! `CopilotSession`/`CopilotMessage` aggregates, `ToolExecution`, capability
-//! routing, and LLM orchestration remain in the binary (coupled to HTTP and the
-//! inference engines) and migrate in later phases.
+//! The offline AI labeling copilot: chat, tool calling, AI actions, and the
+//! approval workflow. [`contracts`] holds the command request/response DTOs (the
+//! chat-turn payload, the approved-action payload, and the connection-test
+//! payload/result). [`domain`] holds the pure decision logic the copilot is
+//! built on — the deterministic intent router, the LLM plan parser/validator,
+//! the QA-diff engine, label-suggestion parsing, and the [`CopilotLlmConfig`].
+//! It is pure: no HTTP, no Tauri, no inference engines. Tool execution + LLM and
+//! inference orchestration move into the application layer (later phases),
+//! depending on ports the binary implements.
+//!
+//! [`CopilotLlmConfig`]: domain::CopilotLlmConfig
 
 pub mod contracts;
+pub mod domain;
